changed most unsigned types to ints, big cleanups, cmake threads linking improvements, IPO/LTO check support

+++ src/Board.h
@@ -35,7 +35,7 @@ public:
     }
 
 #pragma region ===== Getters =====
-    [[nodiscard]] inline size_t HistorySize() const { return history_cnt; }
+    [[nodiscard]] inline int HistorySize() const { return history_cnt; }
     [[nodiscard]] inline Score GetEval() const { return state.evaluator.GetBoardEval(); };
     [[nodiscard]] inline Score GetMoveEval(Move m) const { return state.evaluator.GetMoveEval(*this, m); };
     [[nodiscard]] inline Bitboard GetPieces(PieceType pt, Color c) const { return type_bbs[pt] & color_bbs[c]; }
@@ -103,8 +103,8 @@ private:
 
     BoardState state;
     BoardState history[MAX_GAME_LENGTH];
-    size_t history_cnt;
-    size_t uci_moves_cnt;
+    int history_cnt;
+    int uci_moves_cnt;
 
     // original positions of rooks that are available for castling
     Bitboard origin_rooks[COLOR_NR][CS_NR];
